+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Added
+
+- Extensions for Windows release artifacts
+
 ## 7.0.0 - 2024-09-17
 
 ### Added