+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## Unreleased
 
+### Changed
+
+- Use `suppression_reason` instead of deprecated `suppressed` field
+
 ## 7.1.0 - 2024-09-24
 
 ### Added