+++ CHANGELOG.md
@@ -12,6 +12,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Use `suppression_reason` instead of deprecated `suppressed` field
 
+### Fixed
+
+- Package subcommand failing to parse API responses
+
 ## 7.1.0 - 2024-09-24
 
 ### Added
@@ -21,7 +25,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Phylum project file paths on Windows
-- Package subcommand failing to parse API responses
 
 ## 7.0.0 - 2024-09-17